Speed up cell<->hit linking.
The cell<->hit linking in CreatePositionedCaloCells was N^2 in the
number of cells.  When noise was enabled for ALLEGRO, this was thus
taking an excessive about of time.

Keep a multimap from cellid->hits to speed this up; this changes the
behavior from N^2 to linear.  For an example ALLEGRO reco digi+reco job
with noise enabled, this reduces the time from ~19m to ~3m.